Seperate Chromium and Edge UI's
Seperate the chromium and EDGE UI's into lorca.NewChromium and
lorca.NewEdge so that the user can choose to use Edge instead of
chrome when they have booth installed. Also, do not automatically
add in all the default arguments. Instead make these public through
AdditionalEdgeArgs and AdditionalChromiumArgs so that the user is not
forced to use all the default arguments.